requirements.txt
action-tutorials-py==0.20.4
albucore==0.0.10
albumentations==1.4.8
ament-clang-format==0.12.11
ament-copyright==0.12.11
ament-cppcheck==0.12.11
ament-cpplint==0.12.11
ament-flake8==0.12.11
ament-index-python==1.4.0
ament-lint==0.12.11
ament-lint-cmake==0.12.11
ament-package==0.14.0
ament-pep257==0.12.11
ament-uncrustify==0.12.11
ament-xmllint==0.12.11
annotated-types==0.7.0
appdirs==1.4.4
argcomplete==1.8.1
attrs==21.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.4
bloom==0.12.0
breezy==3.2.1
Brotli==1.0.9
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
certifi==2020.6.20
chardet==4.0.0
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.28
colcon-common-extensions==0.3.0
colcon-core==0.16.1
colcon-defaults==0.2.8
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-mixin==0.2.3
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.8
colcon-recursive-crawl==0.2.3
colcon-ros==0.4.1
colcon-test-result==0.3.8
colcon-zsh==0.5.0
configobj==5.0.6
cov-core==1.15.0
coverage==6.2
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
demo-nodes-py==0.20.4
Deprecated==1.2.13
distlib==0.3.4
distro==1.7.0
docutils==0.17.1
domain-coordinator==0.10.0
dulwich==0.20.31
empy==3.3.4
examples-rclpy-executors==0.15.1
examples-rclpy-minimal-action-client==0.15.1
examples-rclpy-minimal-action-server==0.15.1
examples-rclpy-minimal-client==0.15.1
examples-rclpy-minimal-publisher==0.15.1
examples-rclpy-minimal-service==0.15.1
examples-rclpy-minimal-subscriber==0.15.1
fastbencode==0.0.5
fastimport==0.9.14
filelock==3.15.1
flake8==4.0.1
fonttools==4.29.1
fs==2.4.12
fsspec==2024.6.0
gast==0.5.2
gpg==1.16.0
html5lib==1.1
httplib2==0.20.2
huggingface-hub==0.23.4
idna==3.3
imageio==2.34.1
importlib-metadata==4.6.4
iniconfig==1.1.1
jeepney==0.7.1
Jinja2==3.1.4
joblib==1.4.2
keyring==23.5.0
kiwisolver==1.3.2
lark==1.1.1
launch==1.0.6
launch-ros==0.19.7
launch-testing==1.0.6
launch-testing-ros==0.19.7
launch-xml==1.0.6
launch-yaml==1.0.6
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
lxml==4.8.0
lz4==3.1.3+dfsg
MarkupSafe==2.1.5
matplotlib==3.5.1
mccabe==0.6.1
mercurial==6.1.1
more-itertools==8.10.0
mpi4py==3.1.3
mpmath==0.0.0
netifaces==0.11.0
networkx==3.3
nose2==0.9.2
notify2==0.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
olefile==0.46
opencv-python==4.10.0.82
opencv-python-headless==4.10.0.82
osrf-pycommon==2.0.2
packaging==21.3
patiencediff==0.2.1
pillow==10.3.0
pluggy==0.13.0
ply==3.11
psutil==5.9.0
py==1.10.0
pycairo==1.20.1
pycodestyle==2.8.0
pydantic==2.7.4
pydantic_core==2.18.4
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.4.0
PyGithub==1.55
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pytest==6.2.5
pytest-cov==3.0.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-gitlab==2.10.1
pythran==0.10.0
pytz==2022.1
PyYAML==5.4.1
quality-of-service-demo-py==0.20.4
requests==2.25.1
requests-toolbelt==0.9.1
roman==3.3
ros2action==0.18.10
ros2bag==0.15.11
ros2cli==0.18.10
ros2component==0.18.10
ros2doctor==0.18.10
ros2interface==0.18.10
ros2launch==0.19.7
ros2lifecycle==0.18.10
ros2multicast==0.18.10
ros2node==0.18.10
ros2param==0.18.10
ros2pkg==0.18.10
ros2run==0.18.10
ros2service==0.18.10
ros2topic==0.18.10
rosdep==0.24.0
rosdep-modules==0.24.0
rosdistro==0.9.1
rosdistro-modules==0.9.1
rosidl-cli==3.1.5
rosidl-runtime-py==0.9.3
rospkg-modules==1.5.1
rpyutils==0.2.1
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.7
rqt-gui-py==1.1.7
rqt-msg==1.2.0
rqt-plot==1.1.2
rqt-publisher==1.5.0
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
safetensors==0.4.3
scikit-image==0.23.2
scikit-learn==1.5.0
scipy==1.13.1
SecretStorage==3.3.1
sensor-msgs-py==4.2.3
six==1.16.0
snowballstemmer==2.2.0
soupsieve==2.3.1
sros2==0.10.5
sympy==1.9
teleop-twist-keyboard==2.4.0
tf2-ros-py==0.25.6
tf2-tools==0.25.6
threadpoolctl==3.5.0
tifffile==2024.5.22
timm==1.0.3
toml==0.10.2
tomli==2.0.1
topic-monitor==0.20.4
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transforms3d==0.4.1
triton==2.3.1
typing_extensions==4.12.2
ufoLib2==0.13.1
unicodedata2==14.0.0
urllib3==1.26.5
vcstool==0.3.0
vcstools==0.1.42
wadllib==1.3.6
webencodings==0.5.1
wrapt==1.13.3
zipp==1.0.0
